+++ ElvUI_LuckyoneUI/Core/Themes.lua
@@ -60,7 +60,7 @@ function L1UI:RaidFrames(style)
 			["fontOutline"] = "OUTLINE",
 			["justifyH"] = "CENTER",
 			["size"] = 14,
-			["text_format"] = "[namecolor][name:medium]",
+			["text_format"] = "[classcolor][name:medium]",
 			["xOffset"] = 0,
 			["yOffset"] = 0
 		}
@@ -124,7 +124,7 @@ function L1UI:RaidFrames(style)
 			["fontOutline"] = "OUTLINE",
 			["justifyH"] = "CENTER",
 			["size"] = 12,
-			["text_format"] = "[namecolor][name:short]",
+			["text_format"] = "[classcolor][name:short]",
 			["xOffset"] = 0,
 			["yOffset"] = 0
 		}
@@ -197,10 +197,10 @@ function L1UI:SwapTags(group)
 
 		if L1UI.Retail then
 			E.db["unitframe"]["units"]["boss"]["customTexts"]["Luckyone_HP"]["text_format"] = "[perhp<%]"
-			E.db["unitframe"]["units"]["boss"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:last]"
+			E.db["unitframe"]["units"]["boss"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:last]"
 			E.db["unitframe"]["units"]["boss"]["customTexts"]["Luckyone_Power"]["text_format"] = "[powercolor][perpp<%]"
 			E.db["unitframe"]["units"]["pet"]["customTexts"]["Luckyone_HP"]["text_format"] = "[perhp<%]"
-			E.db["unitframe"]["units"]["pet"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor]Pet"
+			E.db["unitframe"]["units"]["pet"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor]Pet"
 		end
 
 		if not L1UI.Retail then
@@ -210,14 +210,14 @@ function L1UI:SwapTags(group)
 
 		if not L1UI.Classic then
 			E.db["unitframe"]["units"]["arena"]["customTexts"]["Luckyone_HP"]["text_format"] = "[perhp<%]"
-			E.db["unitframe"]["units"]["arena"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:medium]"
+			E.db["unitframe"]["units"]["arena"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:medium]"
 			E.db["unitframe"]["units"]["arena"]["customTexts"]["Luckyone_Power"]["text_format"] = "[powercolor][perpp<%]"
 			E.db["unitframe"]["units"]["focus"]["customTexts"]["Luckyone_HP"]["text_format"] = "[perhp<%] | [health:current:shortvalue]"
-			E.db["unitframe"]["units"]["focus"]["customTexts"]["Luckyone_Level"]["text_format"] = "[namecolor][level]"
-			E.db["unitframe"]["units"]["focus"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:last]"
+			E.db["unitframe"]["units"]["focus"]["customTexts"]["Luckyone_Level"]["text_format"] = "[classcolor][level]"
+			E.db["unitframe"]["units"]["focus"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:last]"
 			E.db["unitframe"]["units"]["focus"]["customTexts"]["Luckyone_Power"]["text_format"] = "[powercolor][perpp<%]"
 			E.db["unitframe"]["units"]["focustarget"]["customTexts"]["Luckyone_HP"]["text_format"] = "[perhp<%]"
-			E.db["unitframe"]["units"]["focustarget"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:last]"
+			E.db["unitframe"]["units"]["focustarget"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:last]"
 		end
 
 		E.db["nameplates"]["units"]["ENEMY_NPC"]["health"]["text"]["format"] = "[perhp<%]"
@@ -231,46 +231,46 @@ function L1UI:SwapTags(group)
 		E.db["nameplates"]["units"]["FRIENDLY_NPC"]["level"]["format"] = "[level]"
 		E.db["nameplates"]["units"]["FRIENDLY_PLAYER"]["health"]["text"]["format"] = "[perhp<%]"
 		E.db["nameplates"]["units"]["FRIENDLY_PLAYER"]["level"]["format"] = "[level]"
-		E.db["unitframe"]["units"]["party"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:short]"
+		E.db["unitframe"]["units"]["party"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:short]"
 		E.db["unitframe"]["units"]["player"]["customTexts"]["Luckyone_HP"]["text_format"] = "[health:current:shortvalue] | [perhp<%]"
-		E.db["unitframe"]["units"]["player"]["customTexts"]["Luckyone_Level"]["text_format"] = "[namecolor][level]"
-		E.db["unitframe"]["units"]["player"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:veryshort]"
+		E.db["unitframe"]["units"]["player"]["customTexts"]["Luckyone_Level"]["text_format"] = "[classcolor][level]"
+		E.db["unitframe"]["units"]["player"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:veryshort]"
 		E.db["unitframe"]["units"]["player"]["customTexts"]["Luckyone_Power"]["text_format"] = "[powercolor][perpp<%]"
-		E.db["unitframe"]["units"]["raid"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:short]"
+		E.db["unitframe"]["units"]["raid"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:short]"
 		E.db["unitframe"]["units"]["target"]["customTexts"]["Luckyone_HP"]["text_format"] = "[perhp<%] | [health:current:shortvalue]"
-		E.db["unitframe"]["units"]["target"]["customTexts"]["Luckyone_Level"]["text_format"] = "[namecolor][level]"
-		E.db["unitframe"]["units"]["target"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:last]"
+		E.db["unitframe"]["units"]["target"]["customTexts"]["Luckyone_Level"]["text_format"] = "[classcolor][level]"
+		E.db["unitframe"]["units"]["target"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:last]"
 		E.db["unitframe"]["units"]["target"]["customTexts"]["Luckyone_Power"]["text_format"] = "[powercolor][perpp<%]"
 		E.db["unitframe"]["units"]["targettarget"]["customTexts"]["Luckyone_HP"]["text_format"] = "[perhp<%]"
-		E.db["unitframe"]["units"]["targettarget"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:last]"
+		E.db["unitframe"]["units"]["targettarget"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:last]"
 
 	elseif group == 'luckyone' then
 
 		if L1UI.Retail then
 			E.db["unitframe"]["units"]["boss"]["customTexts"]["Luckyone_HP"]["text_format"] = "[luckyone:health:percent]"
-			E.db["unitframe"]["units"]["boss"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:last]"
+			E.db["unitframe"]["units"]["boss"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:last]"
 			E.db["unitframe"]["units"]["boss"]["customTexts"]["Luckyone_Power"]["text_format"] = "[powercolor][perpp<%]"
-			E.db["unitframe"]["units"]["party"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:short][ >luckyone:healermana:percent]"
+			E.db["unitframe"]["units"]["party"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:short][ >luckyone:healermana:percent]"
 			E.db["unitframe"]["units"]["pet"]["customTexts"]["Luckyone_HP"]["text_format"] = "[luckyone:health:percent]"
-			E.db["unitframe"]["units"]["pet"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor]Pet"
+			E.db["unitframe"]["units"]["pet"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor]Pet"
 		end
 
 		if not L1UI.Retail then
-			E.db["unitframe"]["units"]["party"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:short]"
+			E.db["unitframe"]["units"]["party"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:short]"
 			E.db["unitframe"]["units"]["pet"]["customTexts"]["Luckyone_HP"]["text_format"] = "[luckyone:health:percent]"
 			E.db["unitframe"]["units"]["pet"]["customTexts"]["Luckyone_Name"]["text_format"] = "[happiness:color]Pet[ >happiness:full]"
 		end
 
 		if not L1UI.Classic then
 			E.db["unitframe"]["units"]["arena"]["customTexts"]["Luckyone_HP"]["text_format"] = "[luckyone:health:percent]"
-			E.db["unitframe"]["units"]["arena"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:medium]"
+			E.db["unitframe"]["units"]["arena"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:medium]"
 			E.db["unitframe"]["units"]["arena"]["customTexts"]["Luckyone_Power"]["text_format"] = "[powercolor][perpp<%]"
 			E.db["unitframe"]["units"]["focus"]["customTexts"]["Luckyone_HP"]["text_format"] = "[luckyone:health:percent] | [health:current:shortvalue]"
-			E.db["unitframe"]["units"]["focus"]["customTexts"]["Luckyone_Level"]["text_format"] = "[namecolor][level]"
-			E.db["unitframe"]["units"]["focus"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:last]"
+			E.db["unitframe"]["units"]["focus"]["customTexts"]["Luckyone_Level"]["text_format"] = "[classcolor][level]"
+			E.db["unitframe"]["units"]["focus"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:last]"
 			E.db["unitframe"]["units"]["focus"]["customTexts"]["Luckyone_Power"]["text_format"] = "[powercolor][perpp<%]"
 			E.db["unitframe"]["units"]["focustarget"]["customTexts"]["Luckyone_HP"]["text_format"] = "[luckyone:health:percent]"
-			E.db["unitframe"]["units"]["focustarget"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:last]"
+			E.db["unitframe"]["units"]["focustarget"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:last]"
 		end
 
 		E.db["nameplates"]["units"]["ENEMY_NPC"]["health"]["text"]["format"] = "[luckyone:health:percent]"
@@ -285,16 +285,16 @@ function L1UI:SwapTags(group)
 		E.db["nameplates"]["units"]["FRIENDLY_PLAYER"]["health"]["text"]["format"] = "[luckyone:health:percent]"
 		E.db["nameplates"]["units"]["FRIENDLY_PLAYER"]["level"]["format"] = "[level]"
 		E.db["unitframe"]["units"]["player"]["customTexts"]["Luckyone_HP"]["text_format"] = "[health:current:shortvalue] | [luckyone:health:percent]"
-		E.db["unitframe"]["units"]["player"]["customTexts"]["Luckyone_Level"]["text_format"] = "[namecolor][level]"
-		E.db["unitframe"]["units"]["player"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:veryshort]"
+		E.db["unitframe"]["units"]["player"]["customTexts"]["Luckyone_Level"]["text_format"] = "[classcolor][level]"
+		E.db["unitframe"]["units"]["player"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:veryshort]"
 		E.db["unitframe"]["units"]["player"]["customTexts"]["Luckyone_Power"]["text_format"] = "[powercolor][perpp<%]"
-		E.db["unitframe"]["units"]["raid"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:short]"
+		E.db["unitframe"]["units"]["raid"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:short]"
 		E.db["unitframe"]["units"]["target"]["customTexts"]["Luckyone_HP"]["text_format"] = "[luckyone:health:percent] | [health:current:shortvalue]"
-		E.db["unitframe"]["units"]["target"]["customTexts"]["Luckyone_Level"]["text_format"] = "[namecolor][level]"
-		E.db["unitframe"]["units"]["target"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:last]"
+		E.db["unitframe"]["units"]["target"]["customTexts"]["Luckyone_Level"]["text_format"] = "[classcolor][level]"
+		E.db["unitframe"]["units"]["target"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:last]"
 		E.db["unitframe"]["units"]["target"]["customTexts"]["Luckyone_Power"]["text_format"] = "[powercolor][perpp<%]"
 		E.db["unitframe"]["units"]["targettarget"]["customTexts"]["Luckyone_HP"]["text_format"] = "[luckyone:health:percent]"
-		E.db["unitframe"]["units"]["targettarget"]["customTexts"]["Luckyone_Name"]["text_format"] = "[namecolor][name:last]"
+		E.db["unitframe"]["units"]["targettarget"]["customTexts"]["Luckyone_Name"]["text_format"] = "[classcolor][name:last]"
 
 	end